feat(i64): I64-S1 Phase A — chain-wide Q16 i32→i64 unification
Widen the consensus Q16.16 fixed-point type from an i32 backing to i64
across EVERY precompile family, the learning daemon, and the zkp circuit,
so on-chain Q16 agrees bit-for-bit with the Tier-1-audited
citrate-fed-types kernel. Same fractional resolution (2^-16); only adds
integer headroom (±32_768 → ±1.407e14). Part of the deterministic
state-reset reroll — i64 lives in the node binary/precompiles/genesis,
NOT in CREATE2 bytecode, so all existing contract addresses are preserved.

Arithmetic core (WP-A1/A4):
- q16/mod.rs: Q16(i64); i128 mul/div intermediates; i64 saturation bounds.
- q16/exp.rs: saturation gate moved (real 11→34); exp(11)/exp(16) now
  representable; internal shift bounds i32→i64.
- ops.rs unchanged (Q16-abstract).

Wire formats — full adversarial-fuzz surface (WP-A2/A3/A7):
- belnap 0x0110 + routing 0x0111: 4→8 byte Q16 on the wire; routing
  ARCH_VERSION 1→2. Existing fuzz targets + 50k/100k deterministic sweeps
  + proptests are format-agnostic and re-prove panic-free/deterministic.
- tensor family 0x010A–0x010F (tensor_format/compute): 8-byte i64 element;
  REMOVED the interim silent i32 saturating-clamp (was a corruption risk).

zkp halo2 LinearChip (WP-A8): witness path widened to i128 product +
i64 field recovery (i128_to_halo2_fr, halo2_fr_to_signed_i64). The gates
are field-exact algebraic relations with no bit-range gates, so the
constraint system — and the verifying key — is byte-identical;
CIRCUIT_VERSION_LINEAR_Q16 intentionally stays 1 (owner-confirmed; v4/v5
reserved). New i64-range prove/verify round-trips that would have wrapped
under the old i32 recovery. Full halo2-substrate suite green (622).

Daemon (WP-A4): Q16Weights Vec<i32>→Vec<i64>; aggregator encode_belnap_input
widened to match 0x0110. Pinned-weights IPFS CID changes (expected).

Determinism + formal:
- Cross-platform goldens re-frozen by inspection (exp(11)=3923968000,
  exp(16)=582362333184 — now representable; the latter genuinely needs i64).
- TLC re-run green on the 3 width/version-abstract specs (0 violations);
  VERIFICATION_REPORT I64-S1 addendum.

Kernel parity (WP-A1 acceptance): new EXCLUDED crate core/federated proves
chain Q16 == citrate_fed_types::Q16 (pinned dbbfbff) bit-for-bit over a
15-value edge vector (from_raw/add/sub/mul/div incl. div-by-zero + consts).
Excluded from the workspace because the kernel MSRV (1.96) is ahead of the
chain stable (1.93) and it pulls a private git dep — consensus
`cargo build --workspace` is untouched. Run:
  cargo +1.96.0 test --manifest-path core/federated/Cargo.toml

Verification: execution lib 525, halo2 622, daemon 84 green; workspace
builds clean on 1.93. (Cargo.lock: incidental ed25519-dalek resolver sync.)
Pre-existing/unrelated: 2 ai_opcode_e2e address-byte tests (addr[17] vs
[18]) fail — not touched here.

+++ core/learning-daemon/tests/chaos.rs
@@ -35,7 +35,7 @@ use citrate_learning_daemon::trainer::{
 use ethereum_types::H160;
 use tempfile::TempDir;
 
-fn make_entry(submitter: u8, value_q16: i32) -> EmbeddingEntry {
+fn make_entry(submitter: u8, value_q16: i64) -> EmbeddingEntry {
     EmbeddingEntry {
         submitter: H160::repeat_byte(submitter),
         embedding: vec![value_q16],
